chore: move resource blocks tests to own file
